Improve support for X11
As tested on my bspwm setup, tid works well and as expected for a system
bar. The key changes are that the PhysicalSize is used to determine the
window size as well as the draw texture. Before this commit, the window
size was set as a LogicalSize, which can cause a mismatch between the
surface that is considered to be part of the window and the pixels that
are actually drawn on that surface. This would cause the window to be
strangely large and also messed up positioning.