Let make_window_sticky_on_mac() take a mutable reference to Window
Rather than taking a shared reference (`&Window`), taking a mutable
reference (`&mut Window`) better conveys the intention, here, since this
function mutates something internal to the provided `Window`.

Since we use `as &dyn Trait` aliasing here, and since the mutation
occurs through a pointer to the `NSWindow` object by means of an
`unsafe` transmute, rust's aliasing rules (viz. a shared reference `&T`
cannot be promoted to a `&mut T`) are not checked. But, we can point out
that it is suspect to mutate something behind a shared reference.